Commented code and some edit
