--- src/main-ecm.py
@@ -1,4 +1,0 @@
-from Eye_Controlled_Mouse import ECM
-
-x=ECM
-x.main()
